\docType{package}
\name{VICCbiostat-package}
\alias{VICCbiostat-package}
\alias{VICCbiostat}
\title{VICCbiostat}
\description{
Collection of useful functions for the Vanderbilt-Ingram Cancer Center's Biostatistics Shared Resource and Department of Biostatistics' Division of Cancer Biostatistics.
}
\author{
Will Gray\cr
Department of Biostatistics\cr
Vanderbilt University\cr
\email{w.gray@vanderbilt.edu}
}
\keyword{package}
